refactor and new tests complete
