+++ services/app-web/src/components/FileUpload/FileItemList/FileItemsList.test.tsx
@@ -2,8 +2,9 @@ import React from 'react'
 import { render, screen } from '@testing-library/react'
 import renderer from 'react-test-renderer'
 import userEvent from '@testing-library/user-event'
+import prettyDOM from '@testing-library/react'
 
-import { FileItemT } from '../FileItem/FileItem'
+import { FileItemT } from '../FileProcessor/FileProcessor'
 import { FileItemsList } from './FileItemsList'
 import { TEST_PDF_FILE } from '../../../testHelpers/jestHelpers'
 
@@ -65,7 +66,7 @@ describe('FileItemList component', () => {
     }
 
     beforeEach(() => jest.clearAllMocks())
-    it('renders without errors', () => {
+    it('renders a list without errors', () => {
         const fileItems = [pending, uploadError]
         render(
             <FileItemsList
@@ -76,6 +77,21 @@ describe('FileItemList component', () => {
         )
 
         expect(screen.getAllByRole('listitem').length).toEqual(fileItems.length)
+        expect(screen.getByText(/testFile.pdf/)).toBeInTheDocument()
+        expect(screen.getByText('testFile2.pdf')).toBeInTheDocument()
+    })
+
+    it('renders a table without errors', () => {
+        const fileItems = [pending, uploadError]
+        render(
+            <FileItemsList
+                renderMode="table"
+                fileItems={fileItems}
+                {...buttonActionProps}
+            />
+        )
+        // the table has a header row so we need to add 1 to the length
+        expect(screen.getAllByRole('row').length).toEqual(fileItems.length + 1)
         expect(screen.getByText(/testFile.pdf/)).toBeInTheDocument()
         expect(screen.getByText('testFile2.pdf')).toBeInTheDocument()
     })
@@ -100,7 +116,7 @@ describe('FileItemList component', () => {
         expect(tree).toMatchSnapshot()
     })
 
-    it('button actions work as expected', () => {
+    it('button actions in a list work as expected', () => {
         const fileItems = [uploadError]
         render(
             <FileItemsList
@@ -117,7 +133,24 @@ describe('FileItemList component', () => {
         expect(buttonActionProps.deleteItem).toHaveBeenCalled()
     })
 
-    it('displays error styles for items that have errors', () => {
+    it('button actions in a table work as expected', () => {
+        const fileItems = [uploadError]
+        render(
+            <FileItemsList
+                renderMode="table"
+                fileItems={fileItems}
+                {...buttonActionProps}
+            />
+        )
+
+        userEvent.click(screen.getByText('Retry'))
+        expect(buttonActionProps.retryItem).toHaveBeenCalled()
+
+        userEvent.click(screen.getByText('Remove'))
+        expect(buttonActionProps.deleteItem).toHaveBeenCalled()
+    })
+
+    it('displays error styles for list items that have errors', () => {
         const fileItems = [
             pending,
             uploadError,
@@ -163,5 +196,41 @@ describe('FileItemList component', () => {
         expect(duplicateErrorListItem).toHaveClass(
             'bg-secondary-lighter border-secondary '
         )
+    })
+
+    it('displays error styles for table rows that have errors', () => {
+        const fileItems = [
+            pending,
+            uploadError,
+            scanningError,
+            complete,
+            duplicateError,
+            scanning,
+        ]
+        render(
+            <FileItemsList
+                renderMode="table"
+                fileItems={fileItems}
+                {...buttonActionProps}
+            />
+        )
+
+        const rows = screen.getAllByRole('row')
+        const loadingRow = rows[1]
+        const uploadErrorRow = rows[2]
+        const scanningErrorRow = rows[3]
+        const completeRow = rows[4]
+        const duplicateErrorRow = rows[5]
+        const scanningRow = rows[6]
+
+        // Items not in error state
+        expect(loadingRow).not.toHaveClass('bg-secondary-lighter')
+        expect(scanningRow).not.toHaveClass('bg-secondary-lighter')
+        expect(completeRow).not.toHaveClass('bg-secondary-lighter')
+
+        // Items in an error state
+        expect(uploadErrorRow).toHaveClass('bg-secondary-lighter')
+        expect(scanningErrorRow).toHaveClass('bg-secondary-lighter')
+        expect(duplicateErrorRow).toHaveClass('bg-secondary-lighter')
     })
 })
