+++ services/app-web/src/components/FileUpload/FileUpload.module.scss
@@ -7,6 +7,10 @@
     }
 }
 
+.loadingImage {
+    float: left;
+}
+
 .fileInput {
     max-width: none;
 }
