+++ services/app-web/src/components/FileUpload/FileUpload.test.tsx
@@ -1,12 +1,8 @@
-import {
-    render,
-    waitFor,
-    screen,
-} from '@testing-library/react'
+import { render, waitFor, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 
 import { FileUpload, FileUploadProps, S3FileData } from './FileUpload'
-import { FileItemT } from './FileItem/FileItem'
+import { FileItemT } from './FileProcessor/FileProcessor'
 import { SPACER_GIF } from './constants'
 import {
     TEST_DOC_FILE,
