+++ services/app-web/src/components/FileUpload/TableWrapper/TableWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { FileItem, FileItemT } from '../FileItem/FileItem'
+import { FileItem, FileItemT } from '../FileProcessor/FileProcessor'
 import { Table } from '@trussworks/react-uswds'
 
 type TableWrapperProps = {
